+++ realm/realm-library/src/androidTest/java/io/realm/RealmObjectSchemaTests.java
@@ -88,14 +88,17 @@ public class RealmObjectSchemaTests {
             realm = Realm.getInstance(realmConfig);
             realmSchema = realm.getSchema();
             DOG_SCHEMA = realmSchema.get("Dog");
-            schema = realmSchema.get("NewClass");
+            // Typed realm cannot get NewClass schema since it is not pre-defined.
+            schema = realmSchema.get("Dog");
         }
         realm.beginTransaction();
     }
 
     @After
     public void tearDown() {
-        realm.cancelTransaction();
+        if (realm.isInTransaction()) {
+            realm.cancelTransaction();
+        }
         realm.close();
     }